Move validation to eks subdir
