+++ cdk_infra/lib/utils/eks/validate-config-schema.ts
